models/requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
deap==1.3.1
decorator==4.4.2
idna==2.10
imageio==2.9.0
imutils==0.5.3
isort==5.6.4
joblib==0.17.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.4
Pillow==8.0.1
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn-pandas==2.0.3
stopit==1.1.2
threadpoolctl==2.1.0
tifffile==2020.11.18
toml==0.10.2
tqdm==4.54.1
update-checker==0.18.0
urllib3==1.26.2
wrapt==1.12.1
xgboost==1.3.0.post0
xlrd==1.2.0
